Update first_task.py
changed first_task in functions 1, lab3

+++ lab3/Functions1/first_task.py
@@ -3,7 +3,7 @@
 """
 
 def grams_to_ounces():
-    grams = int(input())
+    grams = float(input())
     ounces = grams / 28.3495231
     print(ounces)
